requirements.txt
certifi==2020.6.20
click==7.1.2
numpy==1.19.4
Pillow==8.0.1
scipy==1.5.4
